requirements-dev.txt
asgiref==3.3.1
blessings==1.7
bpython==0.20.1
certifi==2020.12.5
chardet==3.0.4
curtsies==0.3.4
Django==3.1.4
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.2
greenlet==0.4.17
gunicorn==20.0.4
idna==2.10
Markdown==3.3.3
product==0.1.1.1
Pygments==2.7.3
python-decouple==3.3
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
wcwidth==0.2.5
